Working Versions
Okay,

So I made a few changes to make it more legible, I think your main issue is the organization of your parens, and the uses of fluidPages versus fluidRows.

Reactive - Tomorrow we are going to go through each change in the diff here and I want you to explain to me what's going on in each section of the code.